Fixes DetailManager, log print to VS,, FPS fast update on Server, Removed Textures In GPU Memory

+++ src/Layers/xrRender/ParticleEffect.cpp
@@ -436,7 +436,7 @@ void ParticleRenderStream( LPVOID lpvParams )
 	#endif // _GPA_ENABLED
 
 			float sina = 0.0f , cosa = 0.0f;
-			DWORD angle = 0xFFFFFFFF;
+			float angle = 0xFFFFFFFF;
 
 			PRS_PARAMS* pParams = (PRS_PARAMS *) lpvParams;
